Separated files by folders + compilation fixed. Still raw needs updated tests
